Refactor study store
